# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /plants/v2/:
    delete:
      tags:
      - Plants
      summary: ''
      description: 'Removes plants from a Facility’s inventory while recording the reason for their disposal.</br><b>Permissions Required</b>: View Veg/Flower Plants • Destroy Veg/Flower Plants'
      externalDocs:
        x-doc-key: delete_plants_v2_
      operationId: DeletePlants
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility that intends to destroy the plants.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/additives:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves additive records applied to plants at a specified Facility.</br><b>Permissions Required</b>: View/Manage Plants Additives'
      externalDocs:
        x-doc-key: get_plants_v2_additives
      operationId: GetPlantAdditives
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to retrieve plant additives.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
    post:
      tags:
      - Plants
      summary: ''
      description: 'Records additive usage details applied to specific plants at a Facility.</br><b>Permissions Required</b>: Manage Plants Additives'
      externalDocs:
        x-doc-key: post_plants_v2_additives
      operationId: PostPlantAdditives
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record Plant additives.
        format: ''
      responses:
        200:
          description: Ok
...
  /plants/v2/flowering:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves flowering-phase plants at a specified Facility, optionally filtered by last modified date.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_flowering
      operationId: GetPlantFlowering
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of flowering plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/growthphases:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves the list of growth phases supported by a specified Facility.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_growthphases
      operationId: GetPlantGrowthPhases
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of available Plant Growth Phases.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/{id}:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a Plant by Id.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_{id}
      operationId: GetPlantById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Plant will be validated against the specified License Number. If not specified, the Plant will be validated against all of the User's current Facilities. Please note that if the Plant is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/inactive:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves inactive plants at a specified Facility.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_inactive
      operationId: GetPlantInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/{label}:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a Plant by label.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_{label}
      operationId: GetPlantByLabel
      parameters:
      - type: string
        name: label
        in: path
        required: true
        description: Route parameter label
        format: ''
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Plant will be validated against the specified License Number. If not specified, the Plant will be validated against all of the User's current Facilities. Please note that if the Plant is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/mother:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves mother-phase plants at a specified Facility.</br><b>Permissions Required</b>: View Mother Plants'
      externalDocs:
        x-doc-key: get_plants_v2_mother
      operationId: GetPlantMother
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of mother plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/mother/inactive:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves inactive mother-phase plants at a specified Facility.</br><b>Permissions Required</b>: View Mother Plants'
      externalDocs:
        x-doc-key: get_plants_v2_mother_inactive
      operationId: GetPlantMotherInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive mother plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/mother/onhold:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves mother-phase plants currently marked as on hold at a specified Facility.</br><b>Permissions Required</b>: View Mother Plants'
      externalDocs:
        x-doc-key: get_plants_v2_mother_onhold
      operationId: GetPlantMotherOnHold
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of mother plants on hold.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/onhold:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves plants that are currently on hold at a specified Facility.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_onhold
      operationId: GetPlantOnHold
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of plants on hold.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/vegetative:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves vegetative-phase plants at a specified Facility, optionally filtered by last modified date.</br><b>Permissions Required</b>: View Veg/Flower Plants'
      externalDocs:
        x-doc-key: get_plants_v2_vegetative
      operationId: GetPlantVegetative
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of vegetating plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plants/v2/waste:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a list of recorded plant waste events for a specific Facility.</br><b>Permissions Required</b>: View Plants Waste'
      externalDocs:
        x-doc-key: get_plants_v2_waste
      operationId: GetPlantWaste
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of waste plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
    post:
      tags:
      - Plants
      summary: ''
      description: 'Records waste events for plants at a Facility, including method, reason, and location details.</br><b>Permissions Required</b>: Manage Plants Waste'
      externalDocs:
        x-doc-key: post_plants_v2_waste
      operationId: PostPlantWaste
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record waste.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste/{id}/package:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a list of package records linked to the specified plantWasteId for a given facility.</br><b>Permissions Required</b>: View Plants Waste'
      externalDocs:
        x-doc-key: get_plants_v2_waste_{id}_package
      operationId: GetPlantWasteByIdPackage
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of waste plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste/{id}/plant:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a list of plants records linked to the specified plantWasteId for a given facility.</br><b>Permissions Required</b>: View Plants Waste'
      externalDocs:
        x-doc-key: get_plants_v2_waste_{id}_plant
      operationId: GetPlantWasteByIdPlant
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of waste plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste/methods/all:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retrieves a list of all available plant waste methods for use within a Facility.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_waste_methods_all
      operationId: GetPlantWasteMethodAll
      parameters:
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plants/v2/waste/reasons:
    get:
      tags:
      - Plants
      summary: ''
      description: 'Retriveves available reasons for recording mature plant waste at a specified Facility.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plants_v2_waste_reasons
      operationId: GetPlantWasteReasons
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of waste reasons.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
...